Finish the sign up/in container section
Originally wanted to align the button to the right side but setting the
container to flex created a weird interaction where the butotn would
span across two lines one for 'Sign' and one for 'Up'. Also the button
would grow vertically creating a weirdly large heighted button for some
reason. Couldn't fix it so just avoided flex altogether and used basic
blocks and inline with margin to align it to the bottom of the container.